Fix read >= 64KB
(attention, untested)

+++ utility/SdFat.h
@@ -238,7 +238,7 @@ class SdFile : public Print {
     uint8_t b;
     return read(&b, 1) == 1 ? b : -1;
   }
-  int16_t read(void* buf, uint16_t nbyte);
+  int32_t read(void* buf, size_t nbyte);
   int8_t readDir(dir_t* dir);
   static uint8_t remove(SdFile* dirFile, const char* fileName);
   uint8_t remove(void);
